First attempt at making test files
Makefile should compare the reference and generated versions of output
from CountSNPASE and GetGeneASE, and ensure that they are the same.

I can't tell why, but CountSNPASE is not producing the same output from
run to run.  I am pretty sure I'm setting the seed properly, but it's
not working.  Will get someone else to look at it.