Make random seed results consistent
Setting the seed of the random number generator did not previously
guarantee an identical output as the order of the dictionary of snps
varied. To overcome this, we now sort the dictionary keys prior to
iterating through them. Because there a speed cost to this, we only do
it if the random seed is set.