Use the indexed bam files
My first thought was that it would make sense to pull out only things
that overlap a SNP.  Maybe I'm implementing it wrong, but it seems to be
a LOT slower. Most of this seems to be spent in the pysam specific code:

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   177388   34.370    0.000   34.370    0.000 calignmentfile.pyx:1742(cnext)
        1   11.160   11.160   58.564   58.564 CountSNPASE.py:99(Get_Potential_SNPS_indexed)
        1    9.302    9.302   76.427   76.427 CountSNPASE.py:373(main)
   698145    3.220    0.000    3.220    0.000 calignedsegment.pyx:1328(get_reference_positions)

The above profiling from just the first minute or two, before I killed
it.