First attempt at calculating ASE by assigning reads
Rather than a snp-based approach for the counting, this assigns a read
as reference, alternate, chimeric/misread, or no snp data, then for each
gene just prints the number of reads.

It's surprisingly fast (assuming the sorting has already been done,
otherwise this approach won't work)---it takes only about a minute to do
1.8M reads on a single core.

Multiprocessing doesn't seem to work---pysam doesn't work well with
concurrency, or at least didn't in past versions. It's not obvious to me
that it can't, but it's not working at the moment, so that might be
something to shoot for.