More ASE quantitation functions
Both of these are an attempt to give more credence to more heavily
expressed genes.

The pref_expr is an expression-weighted preference index---divide by
the sqrt of the sum, rather than dividing by the sum, (equivalent to the
preference index times the sqrt(sum)).

diff_expr is just the absolute difference multiplied by the sum of the
allele counts