Command line option for additional fields
David Gokhman [11:57 AM]
Hey, in the gff files that are used as input for the
GetGeneASEbyReads.py script there are some super-useful fields.
Currently only the gene name is extracted, so I was wondering if you
could add the other fields to the output, especially the HGNC, GenBank,
GeneID and product, but actually all fields would be useful. What do you
think?

peter_combs [12:02 PM]
I think that could be doable... Probably the easiest way to do it would
be to pass in the extra fields as command line arguments, and it would
just add them as columns to the table.